% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsem_builder.R
\name{add_group}
\alias{add_group}
\title{Add a SEM group to the visualization}
\usage{
add_group(
  builder,
  name,
  model = NULL,
  object = NULL,
  level = NULL,
  x = 0,
  y = 0,
  width = NULL,
  height = NULL,
  type = NULL
)
}
\arguments{
\item{builder}{A ggsem_builder object}

\item{name}{Group name (required)}

\item{model}{Model object (lavaan, blavaan, etc.)}

\item{object}{Visualization object (qgraph, semPaths, etc.)}

\item{level}{Group level of multi-group model or multi-group data (e.g., 'Pasteur')}

\item{x}{X-coordinate for center position (default: 0)}

\item{y}{Y-coordinate for center position (default: 0)}

\item{width}{Width of visualization area (default: 25)}

\item{height}{Height of visualization area (default: 25)}

\item{type}{Type of analysis: 'sem' or 'network' (auto-detected)}
}
\value{
Updated ggsem_builder object
}
\description{
Add a SEM group to the visualization
}
