% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_annotations.R
\name{draw_annotations}
\alias{draw_annotations}
\title{Write Text Annotations from an Annotation CSV File (from ggsem Shiny app) on a ggplot2 object}
\usage{
draw_annotations(p, annotations_data, zoom_level)
}
\arguments{
\item{p}{A ggplot2 object}

\item{annotations_data}{The object that stores the CSV file containing information about text annotations from the ggsem Shiny app.}

\item{zoom_level}{A numeric value to control the zoom level of the plot. Default is 1.2.}
}
\value{
A ggplot2 object
}
\description{
Write Text Annotations from an Annotation CSV File (from ggsem Shiny app) on a ggplot2 object
}
\examples{
library(ggplot2)

annotations_data <- data.frame(
text = 'Square One', x = 26, y = 300, font = 'Arial',
size = 20, color = '#000000', angle = 0, alpha = 1,
fontface = 'bold', math_expression = FALSE,
lavaan = FALSE
)

p <- ggplot(mtcars) + geom_point(aes(mpg, disp))

p1 <- draw_annotations(p, annotations_data, zoom_level = 1.2)
}
