% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_axis_range.R
\name{get_axis_range}
\alias{get_axis_range}
\title{Get axis range of a ggplot2 object}
\usage{
get_axis_range(plot)
}
\arguments{
\item{plot}{ggplot2 output from csv_to_ggplot()}
}
\value{
A ggplot2 object
}
\description{
A function to calculate the range of x- and y- axes.
}
\examples{
library(ggplot2)
ggplot(mtcars) + geom_point(aes(mpg, disp)) -> p1
get_axis_range(p1)

}
