% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ggsem}
\alias{ggsem}
\title{Create path model plot}
\usage{
ggsem(
  fit,
  filename,
  title = "Path Model",
  layout_df = NA,
  rename_nodes = F,
  cols = c("#E41A1C", "#377EB8", "grey80"),
  new_node_names = NA,
  layout = "auto",
  alpha = 0.05,
  exclude = "none"
)
}
\arguments{
\item{fit}{The lavaan::sem - created object}

\item{title}{Character string with the plot title}

\item{layout_df}{optional data frame with the x and y coordinates for each variable in the plot}

\item{rename_nodes}{will you be supplying a look up table to rename the nodes}

\item{cols}{vector of colors for the arrows}

\item{new_node_names}{Named vector to rename the variables}

\item{layout}{select from a variety of layout algorithms. can be "auto",
"igraph", "manual", "dendrogram", "linear", "matrix", "treemap", "circlepack",
"partition", or "hive". see ?ggraph::create_layout() for more details.}

\item{alpha}{threshold for significance. Defaults to 0.05}

\item{exclude}{variables to exclude from the plot}
}
\description{
Create path model plot
}
\examples{

# library(tidyverse)
# library(lavaan)
# library(gglavaan)
# library(ggraph)

#fake_data <- data.frame(x = runif(100, -10, 10)) \%>\%
#   mutate(noise1 = rnorm(100),
#         y = 2*x + noise1,
#          noise2= rnorm(100),
#          z = x + y + noise2)

#mod <- 'y ~ x
#         z ~ x + y'

#fit <- sem(mod, data = fake_data)

#ggsem1(fit = fit)
}
