% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ggsem1}
\alias{ggsem1}
\title{Create path model plot}
\usage{
ggsem1(
  fit,
  filename,
  title = "Path Model",
  layout_df = NA,
  rename_nodes = F,
  cols = c("#E41A1C", "#377EB8", "grey80"),
  new_node_names = NA,
  layout = "auto",
  alpha = 0.05,
  exclude = "none",
  h2bmask = FALSE
)
}
\description{
Create path model plot
}
