% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsem_launch.R
\name{ggsem_launch.default}
\alias{ggsem_launch.default}
\title{Launch ggsem Shiny application from objects}
\usage{
\method{ggsem_launch}{default}(x, center_x = NULL, center_y = NULL, ...)
}
\arguments{
\item{x}{Any compatible R object (lavaan, igraph, qgraph, etc.)}

\item{center_x}{X-coordinate for center position}

\item{center_y}{Y-coordinate for center position}

\item{...}{Additional arguments passed to Shiny app}
}
\value{
Runs Shiny application
}
\description{
Launch ggsem Shiny application from objects
}
