% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsem_launch.R
\name{ggsem_launch.ggsem_builder}
\alias{ggsem_launch.ggsem_builder}
\title{Launch ggsem Shiny application from builder}
\usage{
\method{ggsem_launch}{ggsem_builder}(x, session = NULL, ...)
}
\arguments{
\item{x}{A ggsem_builder object}

\item{session}{Session type ('sem' or 'network')}

\item{...}{Additional arguments passed to Shiny app}
}
\value{
Runs Shiny application
}
\description{
Launch ggsem Shiny application from builder
}
