% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsem_launch.R
\name{ggsem_launch}
\alias{ggsem_launch}
\title{Launch ggsem Shiny application}
\usage{
ggsem_launch(x = NULL, ...)
}
\arguments{
\item{x}{A ggsem_builder object, model object, visualization object, or NULL}

\item{...}{Additional arguments passed to Shiny app}
}
\value{
Runs Shiny application
}
\description{
Launches the ggsem Shiny application for interactive SEM and network visualization.
}
