% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_legend}
\alias{make_legend}
\title{Make a nice-looking legend}
\usage{
make_legend(
  scols = c("grey80", "#E41A1C", "#377EB8"),
  signs = c("positive", "negative", "neutral")
)
}
\arguments{
\item{scols}{the colors of thes}

\item{signs}{the names associated with the colors}
}
\description{
You can add the legend to a path model plot using ggpubr::ggarrange(), for
example.
}
