% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{random_layout}
\alias{random_layout}
\title{Create a random layout}
\usage{
random_layout(fit)
}
\description{
This is the first step to creating a manual layout. Change the x and y values
as appropriate (using dplyr::mutate(), for example) to make things look just
right.
}
