% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsem_builder.R
\name{set_type}
\alias{set_type}
\title{Set default type for the builder}
\usage{
set_type(builder, type = "sem")
}
\arguments{
\item{builder}{A ggsem_builder object}

\item{type}{Default type: 'sem' or 'network'}
}
\value{
Updated ggsem_builder object
}
\description{
Change the default type for subsequent add_group() calls
}
