% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_logical}
\alias{valid_logical}
\title{Validate logical input}
\usage{
valid_logical(x)
}
\arguments{
\item{x}{A value indicating a logical input.}
}
\value{
A logical value (\code{TRUE} or \code{FALSE}). Defaults to \code{FALSE}.
}
\description{
Validate logical input
}
\keyword{internal}
