% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_ggplot.R
\name{csv_to_ggplot}
\alias{csv_to_ggplot}
\title{Convert CSV files (from ggsem Shiny app) to ggplot2 output}
\usage{
csv_to_ggplot(
  points_data = NULL,
  lines_data = NULL,
  annotations_data = NULL,
  loops_data = NULL,
  element_order = c("lines", "points", "self_loops", "annotations"),
  zoom_level = 1.2,
  horizontal_position = 0,
  vertical_position = 0
)
}
\arguments{
\item{points_data}{The object that stores the CSV file containing information about points from the ggsem Shiny app. The default is NULL.}

\item{lines_data}{The object that stores the CSV file containing information about lines from the ggsem Shiny app. The default is NULL.}

\item{annotations_data}{The object that stores the CSV file containing information about text annotations from the ggsem Shiny app. The default is NULL.}

\item{loops_data}{The object that stores the CSV file containing information about self-loop arrows from the ggsem Shiny app. The default is NULL.}

\item{element_order}{Order of the graphical elements on display. This is the order in which the graphical elements are added. So if it is written later, then it gets added later (more front),
such as: c("lines", "points", "self_loops", "annotations"), which sets annotations to be added last (and hence most front).}

\item{zoom_level}{A numeric value to control the zoom level of the plot. Default is 1.2.}

\item{horizontal_position}{A numeric value for adjusting the horizontal position of the plot. Default is 0.}

\item{vertical_position}{A numeric value for adjusting the vertical position of the plot. Default is 0.}
}
\value{
Returns a ggplot2 object.
}
\description{
This function converts the four CSV files from the ggsem Shiny app into a ggplot2 output object.
The ggplot2 output can then be modified using standard ggplot2 functions, such as ggtitle() and annotate().
}
\examples{
library(ggplot2)

# CSV files from ggsem app
points_data <- read.csv('https://www.smin95.com/points2.csv')
lines_data <- read.csv('https://www.smin95.com/lines2.csv')

csv_to_ggplot(points_data = points_data,
              lines_data = lines_data,
              zoom_level = 1.4, # From the ggsem app
              horizontal_position = 14, # From the ggsem app
              element_order = c('lines', 'points')) # order priority: lines < points


}
