% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_axis_range.R
\name{get_axis_range}
\alias{get_axis_range}
\title{Get axis range of a ggplot object}
\usage{
get_axis_range(plot)
}
\arguments{
\item{plot}{ggplot output from csv_to_ggplot()}
}
\value{
A list object that has two elements, each of which has two vector values. The first
element stores the minimum and maximum values of the plot's x-axis range, while the second
element stores the minimum and maximum values of the plot's y-axis range.
}
\description{
A function to calculate the range of x- and y- axes.
}
\examples{
library(ggplot2)
ggplot(mtcars) + geom_point(aes(mpg, disp)) -> p1
get_axis_range(p1)

}
