% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{ggsem}
\alias{ggsem}
\title{Run ggsem (shiny app) locally through a browser}
\usage{
ggsem()
}
\value{
No return value, called for side effects
}
\description{
Run ggsem (shiny app) locally through a browser
}
