% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsem_builder.R
\name{launch}
\alias{launch}
\title{Launch method for ggsem_builder objects}
\usage{
launch(builder, ...)
}
\arguments{
\item{builder}{A ggsem_builder object}

\item{...}{Additional arguments passed to Shiny app}
}
\value{
Runs Shiny application
}
\description{
Convenience method for builder pattern: builder |> launch()
}
