% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{random_layout}
\alias{random_layout}
\title{Create a random layout}
\usage{
random_layout(fit)
}
\arguments{
\item{fit}{output of lavaan::sem()}
}
\value{
a data frame with columns x and y for positions on the plot, and metric for the name of the variable
}
\description{
This is the first step to creating a manual layout. Change the x and y values
as appropriate (using dplyr::mutate(), for example) to make things look just
right.
}
